move diff display tests to data_model testing file
