+++ test/test_diff.py
@@ -1,7 +1,6 @@
-import sys
 import unittest
 from collections import OrderedDict, namedtuple, deque
-from diffr.data_model import Diff, DiffItem, MappingDiffItem, term
+from diffr.data_model import Diff, DiffItem, MappingDiffItem
 from diffr.patch import patch
 from diffr.diff import (
     _backtrack, _build_lcs_matrix,
@@ -756,138 +755,3 @@ class DiffFunctionTests(unittest.TestCase):
         expected_diff = Diff(str, diffs)
         self.assertEqual(diff_obj, expected_diff)
         self.assertEqual(patch(d1, diff_obj), d2)
-
-
-class DiffStringTests(unittest.TestCase):
-    '''
-    Ultimately the printed diff output will probably be the most useful part of
-    this library. Here we tests the Diff.__str__ correctly represents the Diff.
-    '''
-    # FIXME: As the printable output is so important, every test which creates a
-    # Diff object should probably also test that its str is as expected.
-    def test_diff_item_str(self):
-        item = 'a'
-        di = DiffItem(remove, item)
-        expected_str = remove('{}'.format(item))
-        self.assertEqual(str(di), expected_str)
-
-    def test_mapping_diff_item_str(self):
-        key = 'a'
-        val = [1, 2, 3]
-        di = MappingDiffItem(
-            unchanged, key, insert, val)
-        expected_str = (
-            unchanged('{!s}: '.format(key)) +
-            insert('{!s}'.format(val)))
-        self.assertEqual(str(di), expected_str)
-
-    # ContextBlock tests, the diff output is just a chain of ContextBlocks, the
-    # bulk of output creation is carried out within the ContextBlock.
-    def test_context_banner_is_correct_for_sequences(self):
-        '''
-        Context banners should contain the information you need the two original
-        sequences such that you only get the items contained within the
-        displayed context block.
-        '''
-        seq1 = [0, 1, 2, 3, 0]
-        seq2 = [0, 4, 2, 5, 0]
-        # the useful context for this diff is the slice 1:4 in both sequences
-        s1_start = s2_start = '1'
-        s1_end = s2_end = '4'
-        diff_obj = diff(seq1, seq2)
-        expected_banner = [
-            '@@ {}{},{} {}{},{} @@'.format(
-                remove('-'), remove(s1_start), remove(s1_end),
-                insert('+'), insert(s2_start), insert(s2_end))
-        ]
-        expected_diff_items = [
-            '{} {}'.format(remove('-'), remove('1')),
-            '{} {}'.format(insert('+'), insert('4')),
-            '{} {}'.format(unchanged(' '), unchanged('2')),
-            '{} {}'.format(remove('-'), remove('3')),
-            '{} {}'.format(insert('+'), insert('5'))
-        ]
-        expected_diff_output = '\n'.join(expected_banner + expected_diff_items)
-        # expected_diff_output is unicode type, convert to str for comparison
-        # FIXME: rework assertion without using context blocks
-        # self.assertEqual(
-        #    diff_obj.context_blocks[0].__str__(), str(expected_diff_output))
-
-    def test_no_context_banner_for_non_sequence(self):
-        set1 = {1, 2}
-        set2 = {'a', 'b'}
-        diff_obj = diff(set1, set2)
-        expected_diff_items = [
-            '{} {}'.format(remove('-'), remove('1')),
-            '{} {}'.format(remove('-'), remove('2')),
-            '{} {}'.format(insert('+'), insert('a')),
-            '{} {}'.format(insert('+'), insert('b'))
-        ]
-        # allow the expected output to be unordered
-        # FIXME: rework without using context_blocks
-#        actual_string = diff_obj.context_blocks[0].__str__()
-#        actual_items = actual_string.split('\n')
-#        if sys.version_info.major >= 3:
-#            self.assertCountEqual(expected_diff_items, actual_items)
-#        else:
-#            self.assertItemsEqual(expected_diff_items, actual_items)
-
-    def test_empty_diff(self):
-        set1 = set()
-        set2 = set()
-        diff_obj = diff(set1, set2)
-        expected_diff_output = '{}{}'.format(
-            unchanged('{!s}('.format(type(set1))),
-            unchanged(')'))
-        self.assertEqual(str(diff_obj), expected_diff_output)
-
-    def test_strings_display_on_single_line(self):
-        a = 'this'
-        b = 'that'
-        d = diff(a, b)
-        expected_str = [
-            unchanged('{!s}('.format(type(a))),
-            '@@ {}{},{} {}{},{} @@'.format(
-                remove('-'), remove('0'), remove('4'),
-                insert('+'), insert('0'), insert('4')),
-            ' {}{}{}{}{}{}'.format(
-                unchanged(' '), unchanged(' '), remove('-'), remove('-'),
-                insert('+'), insert('+')),
-            ' {}{}{}{}{}{}'.format(
-                unchanged('t'), unchanged('h'), remove('i'), remove('s'),
-                insert('a'), insert('t')),
-            unchanged(')')
-        ]
-        self.assertEqual(str(d), '\n'.join(expected_str))
-
-    def test_string_diff_wraps_after_term_width(self):
-        a = ''
-        b = 'a' * term.width
-        d = diff(a, b)
-        expected_str = [
-            unchanged('{!s}('.format(type(a))),
-            '@@ {}{},{} {}{},{} @@'.format(
-                remove('-'), remove('0'), remove('0'),
-                insert('+'), insert('0'), insert('{}'.format(term.width))),
-            ' ' + ('{}'.format(insert('+')) * (term.width - 1)),
-            ' ' + ('{}'.format(insert('a')) * (term.width - 1)),
-            ' {}'.format(insert('+')),
-            ' {}'.format(insert('a')),
-            unchanged(')')
-        ]
-        self.assertEqual(str(d), '\n'.join(expected_str))
-
-    def test_string_is_term_width(self):
-        a = ''
-        b = 'a' * (term.width - 1)
-        d = diff(a, b)
-        expected_str = [
-            unchanged('{!s}('.format(type(a))),
-            '@@ {}{},{} {}{},{} @@'.format(
-                remove('-'), remove('0'), remove('0'),
-                insert('+'), insert('0'), insert('{}'.format(term.width - 1))),
-            ' ' + ('{}'.format(insert('+')) * (term.width - 1)),
-            ' ' + ('{}'.format(insert('a')) * (term.width - 1)),
-            unchanged(')')
-        ]
-        self.assertEqual(str(d), '\n'.join(expected_str))
